Started working on a smart diassembler system called CodeDog, which will be monitoring DMA activity, etc. The class will implement full LINQ interfaces to query its disassembly.

+++ Source/Soft64/MipsR4300/Interpreter/BaseInterpreter.cs
@@ -80,7 +80,7 @@ namespace Soft64.MipsR4300.Interpreter
         protected MipsInstruction FetchInstruction(Int64 address)
         {
             m_InstructionBinReader.BaseStream.Position = address;
-            return new MipsInstruction((UInt64)address, m_InstructionBinReader.ReadUInt32());
+            return new MipsInstruction(address, m_InstructionBinReader.ReadUInt32());
         }
     }
 }
